Cleaned up Statistics

+++ src/main/java/org/asourcious/plusbot/Statistics.java
@@ -1,12 +1,11 @@
 package org.asourcious.plusbot;
 
-
 import java.time.OffsetDateTime;
 
 public final class Statistics {
     private Statistics() {}
 
     public static OffsetDateTime startTime;
-    public static long numCommands = 0;
-    public static long numConnections = 0;
+    public static int numCommands = 0;
+    public static int numConnections = 0;
 }
